Move logging to interfaces/http/middlewares
