+++ src/app/index.js
@@ -5,7 +5,6 @@ import constructDomain, { entities } from 'src/domain';
 import constructServer from 'src/interfaces/http/server';
 
 // Infrastructure
-import constructLogger from 'src/infra/logging';
 import constructDb from 'src/infra/db';
 import constructRepositories from 'src/infra/db/repositories';
 import * as encryption from 'src/infra/encryption';
@@ -14,14 +13,13 @@ import * as encryption from 'src/infra/encryption';
 import constructUseCases from 'src/app/use_cases';
 
 const constructApp = (config) => {
-  const logger = constructLogger({ config });
   const db = constructDb({ config });
   const repositories = constructRepositories({ db, entities });
 
   const domain = constructDomain({ encryption });
   const useCases = constructUseCases({ repositories, domain });
 
-  const server = constructServer({ config, logger, useCases });
+  const server = constructServer({ config, useCases });
 
   return {
     start: server.start,
