+++ src/interfaces/http/middlewares/logging/index.js
